requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
nats-py==2.9.0
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
